tests(storybook): mixed layer definition story extracted to misc.examples
